requirements-cpu.txt
torch==2.4.1
torchrec==0.8.0
torchvision==0.18.1
pandas==1.5.3
torchmetrics==1.4.3
iopath==0.1.10
pyre_extensions==0.0.23
mosaicml-streaming==0.9.0
databricks-connect==15.4.2
databricks-vectorsearch